Refactor types for better return types from performance.get* methods on stubs
