+++ package.json
@@ -39,5 +39,5 @@
     "verify": "npm run src && npm run test"
   },
   "types": "./src/index.d.ts",
-  "version": "0.2.2"
+  "version": "0.2.3"
 }
