+++ tslint.json
@@ -8,6 +8,7 @@
         "no-empty": false,
         "no-empty-interface": false,
         "no-shadowed-variable": false,
+        "object-literal-sort-keys": false,
         "typedef": false
     }
 }
